api/requirements.txt
fastapi
uvicorn
pika
python-multipart
psutil
pyna
pydantic
python-dotenv